requirements.txt
eth-keys>=0.4.0
bitcoin-utils>=0.3.0
pillow>=9.0.0
qrcode>=7.3
cairosvg>=2.5.2  # Optional, improves QR code logo rendering
python-bitcoinlib>=0.11.0  # For additional Bitcoin functionality
bitcoin 

# Image processing
opencv-python==4.8.1.78
numpy==1.26.0

# QR code reading
pyzbar==0.1.9

# OCR
pytesseract==0.3.10

# Optional - for handling paths and files
Pillow==10.0.0
